[v0.88][maintainability] Split cli/tests/artifact_builders/learning_runtime.rs by responsibility (Closes #1851)
